-Removed config
