+++ settings.gradle
@@ -6,7 +6,6 @@ include ':UnionPay'
 include ':AmericanExpress'
 include ':LocalPayment'
 include ':PayPal'
-include ':PayPalNative'
 include ':Card'
 include ':BraintreeCore'
 include ':BraintreeDataCollector'
